server/requirements.txt
flask
flask-sqlalchemy
flask-dance[sqla]
flask-cors
python-dotenv
blinker
